backend/requirements.txt
Django==4.1.5
pymongo==4.3.3
djangorestframework==3.14.0